+++ front-page.php
@@ -14,6 +14,12 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly.
 }
 
+/* Force full width content layout */
+add_filter(
+	'genesis_site_layout',
+	'__genesis_return_full_width_content'
+);
+
 /* Remove page title */
 remove_action(
 	'genesis_entry_header',